hikey: gralloc: Add union alias and additional fields to improve interop w/ hikey960 gralloc
Since we use the gralloc_priv.h for hikey960 and hikey, try to
align the variable names a bit using a union. This allows
drm_hwcomposer importer to need less conditionally built code.

Also add logic to provide a byte_stride value, which the
hikey960 gralloc and drm_hwcomposer uses.

Change-Id: I546a02d2f18b2c19a6b06df9135e45963fb6d519
Signed-off-by: John Stultz <john.stultz@linaro.org>